// src/generated/sitemap.ts
export const sitemapEntries: SitemapEntry[] = [
  {
    url: "http://localhost:3000",
    lastModified: "2026-06-29T15:00:39.625Z",
    changeFrequency: "weekly",
    priority: 1,
  },
  {
    url: "http://localhost:3000/tools/developer/base64-encoder",
    lastModified: "2026-06-29T15:00:39.625Z",
    changeFrequency: "weekly",
    priority: 0.8,
  },
  {
    url: "http://localhost:3000/tools/developer/hash-generator",
    lastModified: "2026-06-29T15:00:39.625Z",
    changeFrequency: "weekly",
    priority: 0.8,
  },
  {
    url: "http://localhost:3000/tools/developer/json-formatter",
    lastModified: "2026-06-29T15:00:39.625Z",
    changeFrequency: "weekly",
    priority: 0.8,
  },
  {
    url: "http://localhost:3000/tools/developer/jwt-decoder",
    lastModified: "2026-06-29T15:00:39.625Z",
    changeFrequency: "weekly",
    priority: 0.8,
  },
  {
    url: "http://localhost:3000/tools/developer/url-encoder",
    lastModified: "2026-06-29T15:00:39.625Z",
    changeFrequency: "weekly",
    priority: 0.8,
  },
  {
    url: "http://localhost:3000/tools/developer/uuid-generator",
    lastModified: "2026-06-29T15:00:39.625Z",
    changeFrequency: "weekly",
    priority: 0.8,
  },
  {
    url: "http://localhost:3000/tools/image/image-compress",
    lastModified: "2026-06-29T15:00:39.625Z",
    changeFrequency: "weekly",
    priority: 0.8,
  },
  {
    url: "http://localhost:3000/tools/image/image-crop",
    lastModified: "2026-06-29T15:00:39.625Z",
    changeFrequency: "weekly",
    priority: 0.8,
  },
  {
    url: "http://localhost:3000/tools/image/image-format-convert",
    lastModified: "2026-06-29T15:00:39.625Z",
    changeFrequency: "weekly",
    priority: 0.8,
  },
  {
    url: "http://localhost:3000/tools/image/image-resize",
    lastModified: "2026-06-29T15:00:39.625Z",
    changeFrequency: "weekly",
    priority: 0.8,
  },
  {
    url: "http://localhost:3000/tools/image/image-rotate",
    lastModified: "2026-06-29T15:00:39.625Z",
    changeFrequency: "weekly",
    priority: 0.8,
  },
  {
    url: "http://localhost:3000/tools/pdf/pdf-compress",
    lastModified: "2026-06-29T15:00:39.625Z",
    changeFrequency: "weekly",
    priority: 0.8,
  },
  {
    url: "http://localhost:3000/tools/pdf/pdf-extract-pages",
    lastModified: "2026-06-29T15:00:39.625Z",
    changeFrequency: "weekly",
    priority: 0.8,
  },
  {
    url: "http://localhost:3000/tools/pdf/pdf-merge",
    lastModified: "2026-06-29T15:00:39.625Z",
    changeFrequency: "weekly",
    priority: 0.8,
  },
  {
    url: "http://localhost:3000/tools/pdf/pdf-protect",
    lastModified: "2026-06-29T15:00:39.625Z",
    changeFrequency: "weekly",
    priority: 0.8,
  },
  {
    url: "http://localhost:3000/tools/pdf/pdf-rotate",
    lastModified: "2026-06-29T15:00:39.625Z",
    changeFrequency: "weekly",
    priority: 0.8,
  },
  {
    url: "http://localhost:3000/tools/pdf/pdf-split",
    lastModified: "2026-06-29T15:00:39.625Z",
    changeFrequency: "weekly",
    priority: 0.8,
  },
  {
    url: "http://localhost:3000/tools/pdf/pdf-unlock",
    lastModified: "2026-06-29T15:00:39.625Z",
    changeFrequency: "weekly",
    priority: 0.8,
  },
  {
    url: "http://localhost:3000/tools/text/case-converter",
    lastModified: "2026-06-29T15:00:39.625Z",
    changeFrequency: "weekly",
    priority: 0.8,
  },
  {
    url: "http://localhost:3000/tools/text/remove-duplicate-lines",
    lastModified: "2026-06-29T15:00:39.625Z",
    changeFrequency: "weekly",
    priority: 0.8,
  },
  {
    url: "http://localhost:3000/tools/text/sort-lines",
    lastModified: "2026-06-29T15:00:39.625Z",
    changeFrequency: "weekly",
    priority: 0.8,
  },
  {
    url: "http://localhost:3000/tools/text/text-diff",
    lastModified: "2026-06-29T15:00:39.625Z",
    changeFrequency: "weekly",
    priority: 0.8,
  },
  {
    url: "http://localhost:3000/tools/text/word-counter",
    lastModified: "2026-06-29T15:00:39.625Z",
    changeFrequency: "weekly",
    priority: 0.8,
  },
];